requirements.txt
jsonpickle
demjson3
requests
furl
geojson
jsonpatch
python-dateutil
testcontainers
sqlalchemy
psycopg2
numpy